Introducing PHPStan
Summary:
Ping is currently a bad candidate for introducing PHPStan. It has a ton of errors that
are inherently rooted in the framework and that are difficult to solve.

BUT, using the baseline feature we should at least prevent the code from getting any
worse from here on.

Test Plan: `composer test` should run fine

Reviewers: cesar

Reviewed By: cesar

Differential Revision: https://phabricator.magic3w.com/D637